Remove collision interrupt
Reworks the code to check all sprites for collisions with a bounding box
check instead of using the collision interrupt. This is in preparation
for sprite multiplexing

+++ src/util.h
@@ -34,8 +34,8 @@ struct bb {
 };
 
 struct bb16 {
-    uint16_t north;
-    uint16_t south;
+    uint8_t north;
+    uint8_t south;
     uint16_t east;
     uint16_t west;
 };
